Use latest version of SunVector class. Fix some compilation warnings.

+++ BVPSolverStats.h
@@ -26,7 +26,7 @@ private:
   bool recordStats;
   long numFuncEvals, numSolveIterations, numJacEval;
   long numFuncEvalJac;
-  int numPtsCurrentMesh;
+  size_t numPtsCurrentMesh;
   int numMeshUpdates;
 };
